Refactor matchers out of k8s_test
[nostory]
